requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-tables2==2.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
importlib-metadata==2.0.0
Markdown==3.2.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
zipp==3.2.0
